requirements.txt
graspologic==3.4.1
jsonlines==4.0.0
matplotlib==3.10.0
nano_vectordb==0.0.2
neo4j==5.25.0
nest_asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.61.1
pandas==2.2.3
pydantic==2.10.6
ragas==0.2.2
scikit_learn==1.6.1
seaborn==0.13.2
setuptools==73.0.1
tenacity==9.0.0
tiktoken==0.7.0
tqdm==4.66.5
transformers==4.47.1
umap_learn==0.5.6
xxhash==3.5.0
future==1.0.0
PyYAML==6.0.2
requests==2.32.3
httpx==0.27.2
torch==2.5.1
